feat: TaskBase.CreateTask function

+++ taskutils/rpc/model/task_data_model.go
@@ -7,11 +7,11 @@ type TaskData struct {
 	TaskId           string    `json:"task_id"`
 	TaskType         string    `json:"task_type"`
 	TaskStage        string    `json:"task_stage"`
-	Status           int       `json:"status"`
-	Priority         *int      `json:"priority"`
-	CrtRetryNum      int       `json:"crt_retry_num"`
-	MaxRetryNum      int       `json:"max_retry_num"`
-	MaxRetryInterval int       `json:"max_retry_interval"`
+	Status           int64     `json:"status"`
+	Priority         *int64    `json:"priority"`
+	CrtRetryNum      int64     `json:"crt_retry_num"`
+	MaxRetryNum      int64     `json:"max_retry_num"`
+	MaxRetryInterval int64     `json:"max_retry_interval"`
 	ScheduleLog      string    `json:"schedule_log"`
 	TaskContext      string    `json:"context"`
 	OrderTime        int64     `json:"order_time"`
